refactor: 불필요한 래핑 객체 삭제
