+++ src/main/java/nextstep/sessions/domain/data/vo/OpenInfo.java
@@ -4,13 +4,12 @@ import java.time.LocalDateTime;
 
 public class OpenInfo {
 
-    private CoverImage coverImage;
     private final Duration duration;
 
     public OpenInfo(Duration duration) {
         this.duration = duration;
     }
-    
+
     public LocalDateTime startDate() {
         return duration.startDate();
     }
