+++ src/test/java/nextstep/sessions/infrastructure/CoverImageRepositoryTest.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import nextstep.sessions.domain.data.CoverImage;
 import nextstep.sessions.domain.data.type.ImageType;
-import nextstep.sessions.domain.data.vo.CoverImage;
 import nextstep.sessions.repository.CoverImageRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
